Update System.Text.Json package to version 8.0.6 and increment project version to 2.3.1; refactor MethodWrapper to partial class and extract methods for description and schema creation.
